move out of micropartition
