+++ src/daft-micropartition/src/micropartition.rs
@@ -20,8 +20,6 @@ use daft_scan::storage_config::{NativeStorageConfig, StorageConfig};
 use daft_scan::{ChunkSpec, DataSource, Pushdowns, ScanTask};
 use daft_table::Table;
 
-use futures::stream::BoxStream;
-use futures::{Stream, StreamExt};
 use parquet2::metadata::FileMetaData;
 use snafu::ResultExt;
 
@@ -505,228 +503,6 @@ fn materialize_scan_task(
     Ok((table_values, cast_to_schema))
 }
 
-async fn stream_scan_task(
-    scan_task: Arc<ScanTask>,
-    io_stats: Option<IOStatsRef>,
-    in_order: bool,
-) -> crate::Result<BoxStream<'static, DaftResult<Vec<Table>>>> {
-    let pushdown_columns = scan_task
-        .pushdowns
-        .columns
-        .as_ref()
-        .map(|v| v.iter().map(|s| s.as_str()).collect::<Vec<&str>>());
-    let file_column_names =
-        _get_file_column_names(pushdown_columns.as_deref(), scan_task.partition_spec());
-
-    if scan_task.sources.len() != 1 {
-        return Err(common_error::DaftError::TypeError(
-            "Streaming reads only supported for single source ScanTasks".to_string(),
-        ))
-        .context(DaftCoreComputeSnafu);
-    }
-    let source = scan_task.sources.first().unwrap();
-    let url = source.get_path();
-    let table_values = match scan_task.storage_config.as_ref() {
-        StorageConfig::Native(native_storage_config) => {
-            let io_config = Arc::new(
-                native_storage_config
-                    .io_config
-                    .as_ref()
-                    .cloned()
-                    .unwrap_or_default(),
-            );
-            let multi_threaded_io = native_storage_config.multithreaded_io;
-            let io_client = daft_io::get_io_client_async(multi_threaded_io, io_config)
-                .await
-                .unwrap();
-
-            match scan_task.file_format_config.as_ref() {
-                // ********************
-                // Native Parquet Reads
-                // ********************
-                FileFormatConfig::Parquet(ParquetSourceConfig {
-                    coerce_int96_timestamp_unit,
-                    field_id_mapping,
-                    ..
-                }) => {
-                    let inference_options =
-                        ParquetSchemaInferenceOptions::new(Some(*coerce_int96_timestamp_unit));
-
-                    // TODO: Support iceberg delete files in streaming read
-                    //
-                    // Create vec of all unique delete files in the scan task
-                    // let iceberg_delete_files = scan_task
-                    //     .sources
-                    //     .iter()
-                    //     .flat_map(|s| s.get_iceberg_delete_files())
-                    //     .flatten()
-                    //     .map(String::as_str)
-                    //     .collect::<HashSet<_>>()
-                    //     .into_iter()
-                    //     .collect::<Vec<_>>();
-
-                    // let delete_map = _read_delete_files(
-                    //     iceberg_delete_files.as_slice(),
-                    //     &[url],
-                    //     io_client.clone(),
-                    //     io_stats.clone(),
-                    //     num_parallel_tasks,
-                    //     runtime_handle.clone(),
-                    //     &inference_options,
-                    // )
-                    // .context(DaftCoreComputeSnafu)?;
-
-                    let row_groups = parquet_sources_to_row_groups(&[source.clone()])
-                        .and_then(|rg| rg.first().cloned())
-                        .flatten();
-                    let metadata = scan_task
-                        .sources
-                        .first()
-                        .and_then(|s| s.get_parquet_metadata().cloned());
-                    daft_parquet::read::stream_parquet(
-                        url,
-                        file_column_names.as_deref(),
-                        None,
-                        scan_task.pushdowns.limit,
-                        row_groups,
-                        scan_task.pushdowns.filters.clone(),
-                        io_client.clone(),
-                        io_stats,
-                        &inference_options,
-                        field_id_mapping.clone(),
-                        metadata,
-                        in_order,
-                    )
-                    .await
-                    .context(DaftCoreComputeSnafu)?
-                }
-
-                // ****************
-                // Native CSV Reads
-                // ****************
-                FileFormatConfig::Csv(cfg) => {
-                    let schema_of_file = scan_task.schema.clone();
-                    let col_names = if !cfg.has_headers {
-                        Some(
-                            schema_of_file
-                                .fields
-                                .values()
-                                .map(|f| f.name.as_str())
-                                .collect::<Vec<_>>(),
-                        )
-                    } else {
-                        None
-                    };
-                    let convert_options = CsvConvertOptions::new_internal(
-                        scan_task.pushdowns.limit,
-                        file_column_names
-                            .as_ref()
-                            .map(|cols| cols.iter().map(|col| col.to_string()).collect()),
-                        col_names
-                            .as_ref()
-                            .map(|cols| cols.iter().map(|col| col.to_string()).collect()),
-                        Some(schema_of_file),
-                        scan_task.pushdowns.filters.clone(),
-                    );
-                    let parse_options = CsvParseOptions::new_with_defaults(
-                        cfg.has_headers,
-                        cfg.delimiter,
-                        cfg.double_quote,
-                        cfg.quote,
-                        cfg.allow_variable_columns,
-                        cfg.escape_char,
-                        cfg.comment,
-                    )
-                    .context(DaftCSVSnafu)?;
-                    let read_options =
-                        CsvReadOptions::new_internal(cfg.buffer_size, cfg.chunk_size);
-                    daft_csv::stream_csv(
-                        url.to_string(),
-                        Some(convert_options),
-                        Some(parse_options),
-                        Some(read_options),
-                        io_client.clone(),
-                        io_stats.clone(),
-                        None,
-                    )
-                    .await
-                    .context(DaftCoreComputeSnafu)?
-                }
-
-                // ****************
-                // Native JSON Reads
-                // ****************
-                FileFormatConfig::Json(cfg) => {
-                    let schema_of_file = scan_task.schema.clone();
-                    let convert_options = JsonConvertOptions::new_internal(
-                        scan_task.pushdowns.limit,
-                        file_column_names
-                            .as_ref()
-                            .map(|cols| cols.iter().map(|col| col.to_string()).collect()),
-                        Some(schema_of_file),
-                        scan_task.pushdowns.filters.clone(),
-                    );
-                    // let
-                    let parse_options = JsonParseOptions::new_internal();
-                    let read_options =
-                        JsonReadOptions::new_internal(cfg.buffer_size, cfg.chunk_size);
-
-                    daft_json::read::stream_json(
-                        url.to_string(),
-                        Some(convert_options),
-                        Some(parse_options),
-                        Some(read_options),
-                        io_client,
-                        io_stats,
-                        None,
-                    )
-                    .await
-                    .context(DaftCoreComputeSnafu)?
-                }
-                #[cfg(feature = "python")]
-                FileFormatConfig::Database(_) => {
-                    return Err(common_error::DaftError::TypeError(
-                        "Native reads for Database file format not implemented".to_string(),
-                    ))
-                    .context(DaftCoreComputeSnafu);
-                }
-                #[cfg(feature = "python")]
-                FileFormatConfig::PythonFunction => {
-                    return Err(common_error::DaftError::TypeError(
-                        "Native reads for PythonFunction file format not implemented".to_string(),
-                    ))
-                    .context(DaftCoreComputeSnafu);
-                }
-            }
-        }
-        #[cfg(feature = "python")]
-        StorageConfig::Python(_) => match scan_task.file_format_config.as_ref() {
-            FileFormatConfig::Parquet(ParquetSourceConfig {
-                coerce_int96_timestamp_unit: _,
-                ..
-            }) => todo!("Implement streaming reads for Parquet with Python Storage Config"),
-            FileFormatConfig::Csv(CsvSourceConfig {
-                has_headers: _,
-                delimiter: _,
-                double_quote: _,
-                ..
-            }) => todo!("Implement streaming reads for CSV with Python Storage Config"),
-            FileFormatConfig::Json(_) => {
-                todo!("Implement streaming reads for JSON with Python Storage Config")
-            }
-            FileFormatConfig::Database(daft_scan::file_format::DatabaseSourceConfig {
-                sql: _,
-                conn: _,
-            }) => todo!("Implement streaming reads for Database"),
-            FileFormatConfig::PythonFunction => {
-                todo!("Implement streaming reads for PythonFunction")
-            }
-        },
-    };
-
-    Ok(table_values)
-}
-
 impl MicroPartition {
     /// Create a new "unloaded" MicroPartition using an associated [`ScanTask`]
     ///
@@ -887,69 +663,6 @@ impl MicroPartition {
         }
     }
 
-    pub async fn from_scan_task_streaming(
-        scan_task: Arc<ScanTask>,
-        io_stats: IOStatsRef,
-        morsel_size: usize,
-        in_order: bool,
-    ) -> crate::Result<BoxStream<'static, DaftResult<Arc<Self>>>> {
-        let schema = scan_task.materialized_schema();
-        match (
-            &scan_task.metadata,
-            &scan_task.statistics,
-            scan_task.file_format_config.as_ref(),
-            scan_task.storage_config.as_ref(),
-        ) {
-            // CASE: ScanTask provides all required metadata.
-            // If the scan_task provides metadata (e.g. retrieved from a catalog) we can use it to create an unloaded MicroPartition
-            (Some(metadata), Some(statistics), _, _) if scan_task.pushdowns.filters.is_none() => {
-                let unloaded = Ok(Arc::new(Self::new_unloaded(
-                    scan_task.clone(),
-                    scan_task
-                        .pushdowns
-                        .limit
-                        .map(|limit| TableMetadata {
-                            length: metadata.length.min(limit),
-                        })
-                        .unwrap_or_else(|| metadata.clone()),
-                    statistics.clone(),
-                )));
-                Ok(Box::pin(futures::stream::iter(std::iter::once(unloaded))))
-            }
-
-            // TODO: Enable metadata reads for streaming
-            // CASE: ScanTask does not provide metadata, but the file format supports metadata retrieval
-            // We can perform an eager **metadata** read to create an unloaded MicroPartition
-            // (
-            //     _,
-            //     _,
-            //     FileFormatConfig::Parquet(ParquetSourceConfig {
-            //         coerce_int96_timestamp_unit: _,
-            //         field_id_mapping: _,
-            //         ..
-            //     }),
-            //     StorageConfig::Native(_cfg),
-            // ) => {
-            //     todo!()
-            // }
-
-            // CASE: Last resort fallback option
-            // Perform an eager **data** read
-            _ => {
-                let statistics = scan_task.statistics.clone();
-                let stream = stream_scan_task(scan_task.clone(), Some(io_stats), in_order).await?;
-                let stream = chunk_tables_into_micropartition_stream(
-                    stream,
-                    schema,
-                    scan_task.partition_spec().cloned(),
-                    statistics,
-                    morsel_size,
-                );
-                Ok(Box::pin(stream))
-            }
-        }
-    }
-
     pub fn empty(schema: Option<SchemaRef>) -> Self {
         let schema = schema.unwrap_or(Schema::empty().into());
         Self::new_loaded(schema, Arc::new(vec![]), None)
@@ -1061,37 +774,6 @@ impl MicroPartition {
             Arc::new(tables_with_id),
             self.statistics.clone(),
         ))
-    }
-}
-
-fn chunk_tables_into_micropartition_stream(
-    mut table_stream: BoxStream<'static, DaftResult<Vec<Table>>>,
-    schema: SchemaRef,
-    partition_spec: Option<PartitionSpec>,
-    statistics: Option<TableStatistics>,
-    morsel_size: usize,
-) -> impl Stream<Item = DaftResult<Arc<MicroPartition>>> + Send {
-    async_stream::try_stream! {
-        let mut buffer = vec![];
-        let mut total_rows = 0;
-        while let Some(tables) = table_stream.next().await {
-            let tables = tables?;
-            for table in tables {
-                let casted_table = table.cast_to_schema_with_fill(schema.as_ref(), partition_spec.as_ref().map(|pspec| pspec.to_fill_map()).as_ref())?;
-                total_rows += casted_table.len();
-                buffer.push(casted_table);
-            }
-            if total_rows >= morsel_size {
-                let mp = Arc::new(MicroPartition::new_loaded(schema.clone(), Arc::new(buffer), statistics.clone()));
-                buffer = vec![];
-                total_rows = 0;
-                yield mp;
-            }
-        }
-        if !buffer.is_empty() {
-            let mp = Arc::new(MicroPartition::new_loaded(schema, Arc::new(buffer), statistics));
-            yield mp;
-        }
     }
 }
